+++ packages/defichain-testing/src/index.ts
@@ -1,1 +1,5 @@
-export * from './utils'
+export * from './account'
+export * from './poolpair'
+export * from './rawtx'
+export * from './token'
+export * from './wallet'
